# build as a release build for 10x performance...