--- a/Makefile
+++ b/Makefile
@@ -1,5 +1,5 @@
 run:
-	(cd src; rustc main.rs; ./main; rm main)
+	(cd src; rustc -O main.rs; ./main; rm main)
 
 clean: # rm leftover binary if make run is interrupted
 	rm src/main