haha. obvious find upon first look at any Rust-specific performance tips. well, i hit my goal performance. need to take a closer look for other low hanging fruit

```
case	n	result	dur_ms	dur_pretty
F(2)	2	1	0	570.00ns
F(4)	4	5	0	1.52µs
F(6)	6	35	0	6.46µs
F(8)	8	319	0	40.98µs
F(10)	10	3559	0	350.72µs
F(12)	12	46841	4	4.68ms
F(14)	14	709601	84	84.26ms


F(16)	16	12156445	1295	1.30s
F(18)	18	232349755	25600	25.60s
```